corrected menu issues.
