--- cmd/cmd/web/templates.go
@@ -1,53 +1,0 @@
-package main
-
-import (
-	"html/template"
-	"path/filepath"
-	"time"
-)
-
-func humanDate(t time.Time) string {
-	return t.Format("02 Jan 2006 at 15:04")
-}
-
-type templateData struct {
-	CSRFToken   string
-	CurrentYear int
-	Flash       string
-	Sys         Assumptions
-}
-
-var funcMap = template.FuncMap{
-	"humanDate": humanDate,
-}
-
-func newTemplateCache(dir string) (map[string]*template.Template, error) {
-	cache := map[string]*template.Template{}
-
-	pages, err := filepath.Glob(filepath.Join(dir, "*.page.tmpl"))
-	if err != nil {
-		return nil, err
-	}
-
-	for _, page := range pages {
-		name := filepath.Base(page)
-
-		ts, err := template.New(name).Funcs(funcMap).ParseFiles(page)
-		if err != nil {
-			return nil, err
-		}
-
-		ts, err = ts.ParseGlob(filepath.Join(dir, "*.layout.tmpl"))
-		if err != nil {
-			return nil, err
-		}
-
-		ts, err = ts.ParseGlob(filepath.Join(dir, "*.partial.tmpl"))
-		if err != nil {
-			return nil, err
-		}
-
-		cache[name] = ts
-	}
-	return cache, nil
-}
