--- pkg/pkg/forms/errors.go
@@ -1,17 +1,0 @@
-package forms
-
-type errors map[string][]string
-
-func (e errors) Add(field string, msg string) {
-	e[field] = append(e[field], msg)
-}
-
-func (e errors) Get(field string) string {
-	es := e[field]
-
-	if len(es) == 0 {
-		return ""
-	}
-	return es[0]
-
-}
