--- pkg/pkg/forms/form.go
@@ -1,94 +1,0 @@
-package forms
-
-import (
-  "fmt"
-  "net/url"
-  "regexp"
-  "strings"
-  "unicode/utf8"
-)
-
-// EmailRX email regex
-// TODO: Unused var
-// var EmailRX = regexp.MustCompile("^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$")
-
-// Form struct
-type Form struct {
-  url.Values
-  Errors errors
-}
-
-// New constructor
-// TODO: Unused Function.
-/*
-func New(data url.Values) *Form {
-  return &Form{
-    data,
-    errors(map[string][]string{}),
-  }
-}
-
- */
-
-// Required check required field
-func (f *Form) Required(fields ...string) {
-  for _, field := range fields {
-    value := f.Get(field)
-    if strings.TrimSpace(value) == "" {
-      f.Errors.Add(field, "This field cannot be blank")
-    }
-  }
-}
-
-// MaxLength check inout size
-func (f *Form) MaxLength(field string, d int) {
-  value := f.Get(field)
-  if value == "" {
-    return
-  }
-  if utf8.RuneCountInString(value) > d {
-    f.Errors.Add(field, fmt.Sprintf("This field is too long (maximum is %d characters)", d))
-  }
-}
-
-// PermittedValues verify field is valid value
-func (f *Form) PermittedValues(field string, opts ...string) {
-  value := f.Get(field)
-  if value == "" {
-    return
-  }
-  for _, opt := range opts {
-    if value == opt {
-      return
-    }
-  }
-  f.Errors.Add(field, "This filed is invalid")
-}
-
-// MinLength checks the field is a minimum length
-func (f *Form) MinLength(field string, d int) {
-  value := f.Get(field)
-  if value == "" {
-    return
-  }
-
-  if utf8.RuneCountInString(value) < d {
-    f.Errors.Add(field, fmt.Sprintf("This field is too short (minimum is %d characters)", d))
-  }
-}
-
-// MatchesPattern checks if field matches a regexp pattern
-func (f *Form) MatchesPattern(field string, pattern *regexp.Regexp) {
-  value := f.Get(field)
-  if value == "" {
-    return
-  }
-  if !pattern.MatchString(value) {
-    f.Errors.Add(field, "This field is invalid")
-  }
-}
-
-// Valid check length of form Errors
-func (f *Form) Valid() bool {
-  return len(f.Errors) == 0
-}
